pyproject.toml
[tool.poetry]
name = "halpybot"
version = "2.0.0-alpha"
license = "GPL-3.0-only"
description = "The Hull Seals IRC Utility bot, built with Pydle."
authors = [
    "David Sangrey <rixxan@hullseals.space>",
    "Rik Overveld <rikoverveld@gmail.com>"
]
readme = "README.md"
repository = "https://github.com/HullSeals/HalpyBOT"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
pydle = {extras = ["sasl"], version = "^1.0.1"}
mysqlclient = "^2.2.4"
numpy = "^1.24.4"
boto3 = "^1.34.76"
aiohttp = "^3.9.5"
gitpython = "^3.1.43"
requests = "^2.31.0"
pyperclip = "^1.8.2"
tqdm = "^4.66.2"
attrs = "^23.2.0 "
cattrs = "^23.2.3 "
loguru = "^0.7.0"
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
sqlalchemy = "^2.0.29"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.1"
pendulum = "^3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-httpserver = "^1.0.6"
pytest-asyncio = "^0.20.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.'MESSAGES CONTROL']
extension-pkg-whitelist = "pydantic"
disable = "line-too-long"

[tool.pytest.ini_options]
asyncio_mode = "auto"
